abstract DailyMessageEvent added + evening and morning events refactored
